HSEARCH-3199 Split OrmAutomaticIndexingIT into one test per association type
Because we are going to add more tests, and this test is already massive,
to the point of being barely understandable.
